Exception handling - Delete

+++ src/main/java/com/torres/project/services/UserService.java
@@ -2,8 +2,11 @@ package com.torres.project.services;
 import com.torres.project.entities.User;
 import com.torres.project.repositories.UserRepository;
 
+import com.torres.project.services.exceptions.DatabaseException;
 import com.torres.project.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.*;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.*;
 
 import java.util.List;
@@ -29,7 +32,14 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
